backend/requirements.txt
python-dotenv
cohere
flask